Reorganize project structure and add comments
