--- memory_project/memory_project/asgi.py
@@ -1,8 +1,0 @@
-"""ASGI entry point for asynchronous servers."""
-
-import os
-from django.core.asgi import get_asgi_application
-
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'memory_project.settings')
-
-application = get_asgi_application()
