--- memory_project/memory_project/urls.py
@@ -1,9 +1,0 @@
-"""Project level URL configuration."""
-
-from django.contrib import admin
-from django.urls import path, include
-
-urlpatterns = [
-    path('admin/', admin.site.urls),
-    path('', include('game.urls')),
-]
